move PageStripViewController in own file
